+++ CHANGELOG.md
@@ -8,6 +8,8 @@ glance - what to expact from upgrading to a new version.
 
 ### Added
 
+- Support for community URLs containing Steam ID3.
+
 ### Changed
 
 ### Fixed